fix: skeleton files
